Question1

+++ C++/Question1.cpp
@@ -3,7 +3,7 @@
 using namespace std;
 
 int main(){
-	int num1,num2,prod;
+	float num1,num2,prod;
 	cout<<"Enter first number: ";
 	cin>>num1;
 	cout<<"Enter secund number: ";
